% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_latent_states}
\alias{plot_latent_states}
\title{Plot estimated latent states with credible intervals}
\usage{
plot_latent_states(object, true_states = NULL, ...)
}
\arguments{
\item{object}{A \code{dcVar_sem_fit} object.}

\item{true_states}{Optional T x 2 matrix of true latent states for overlay.}

\item{...}{Additional arguments (unused).}
}
\value{
A ggplot object.
}
\description{
Plot estimated latent states with credible intervals
}
