% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_phi}
\alias{plot_phi}
\title{Plot VAR(1) coefficient matrix as a heatmap}
\usage{
plot_phi(object, var_names = NULL, ...)
}
\arguments{
\item{object}{A fitted model object.}

\item{var_names}{Character vector of variable names for axis labels.}

\item{...}{Additional arguments (unused).}
}
\value{
A ggplot object.
}
\description{
Plot VAR(1) coefficient matrix as a heatmap
}
