% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constant_fit-class.R
\name{print.dcvar_constant_summary}
\alias{print.dcvar_constant_summary}
\title{Print a dcvar_constant_summary object}
\usage{
\method{print}{dcvar_constant_summary}(x, ...)
}
\arguments{
\item{x}{A \code{dcvar_constant_summary} object as returned by
\code{\link[=summary.dcvar_constant_fit]{summary.dcvar_constant_fit()}}.}

\item{...}{Additional arguments (unused).}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print a dcvar_constant_summary object
}
