% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate_fit-class.R
\name{print.dcvar_covariate_summary}
\alias{print.dcvar_covariate_summary}
\title{Print a dcvar_covariate_summary object}
\usage{
\method{print}{dcvar_covariate_summary}(x, ...)
}
\arguments{
\item{x}{A \code{dcvar_covariate_summary} object as returned by \code{summary()}.}

\item{...}{Additional arguments (unused).}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print a dcvar_covariate_summary object
}
