% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmm_fit-class.R
\name{print.dcvar_hmm_summary}
\alias{print.dcvar_hmm_summary}
\title{Print a dcvar_hmm_summary object}
\usage{
\method{print}{dcvar_hmm_summary}(x, ...)
}
\arguments{
\item{x}{A \code{dcvar_hmm_summary} object as returned by
\code{\link[=summary.dcvar_hmm_fit]{summary.dcvar_hmm_fit()}}.}

\item{...}{Additional arguments (unused).}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print a dcvar_hmm_summary object
}
