% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multilevel_fit-class.R
\name{print.dcvar_multilevel_tv_summary}
\alias{print.dcvar_multilevel_tv_summary}
\title{Print a dcvar_multilevel_tv_summary object}
\usage{
\method{print}{dcvar_multilevel_tv_summary}(x, ...)
}
\arguments{
\item{x}{A \code{dcvar_multilevel_tv_summary} object.}

\item{...}{Additional arguments (unused).}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print a dcvar_multilevel_tv_summary object
}
