% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sem_fit-class.R
\name{print.dcVar_sem_summary}
\alias{print.dcVar_sem_summary}
\title{Print a dcVar_sem_summary object}
\usage{
\method{print}{dcVar_sem_summary}(x, ...)
}
\arguments{
\item{x}{A \code{dcVar_sem_summary} object.}

\item{...}{Additional arguments (unused).}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print a dcVar_sem_summary object
}
