% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcvar_fit-class.R
\name{print.dcvar_summary}
\alias{print.dcvar_summary}
\title{Print a dcvar_summary object}
\usage{
\method{print}{dcvar_summary}(x, ...)
}
\arguments{
\item{x}{A \code{dcvar_summary} object as returned by \code{\link[=summary.dcvar_fit]{summary.dcvar_fit()}}.}

\item{...}{Additional arguments (unused).}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print a dcvar_summary object
}
