% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tv_fit-class.R
\name{print.dcvar_tv_summary}
\alias{print.dcvar_tv_summary}
\title{Print a dcvar_tv_summary object}
\usage{
\method{print}{dcvar_tv_summary}(x, ...)
}
\arguments{
\item{x}{A \code{dcvar_tv_summary} object as returned by
\code{\link[=summary.dcvar_tv_fit]{summary.dcvar_tv_fit()}}.}

\item{...}{Additional arguments (unused).}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print a dcvar_tv_summary object
}
