% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectories.R
\name{rho_constant}
\alias{rho_constant}
\title{Generate a constant rho trajectory}
\usage{
rho_constant(n_time, rho = 0.5)
}
\arguments{
\item{n_time}{Number of time points.}

\item{rho}{Constant correlation value (default: 0.5). Must be in [-1, 1].}
}
\value{
Numeric vector of length \code{n_time - 1}.
}
\description{
Generate a constant rho trajectory
}
\examples{
rho_constant(100, rho = 0.5)
}
