% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loo.R
\name{loo.dcvar}
\alias{loo.dcvar}
\alias{loo.dcvar_fit}
\alias{loo.dcvar_hmm_fit}
\alias{loo.dcvar_constant_fit}
\alias{loo.dcvar_multilevel_fit}
\alias{loo.dcvar_sem_fit}
\title{Compute LOO-CV for a fitted model}
\usage{
\method{loo}{dcvar_fit}(x, ...)

\method{loo}{dcvar_hmm_fit}(x, ...)

\method{loo}{dcvar_constant_fit}(x, ...)

\method{loo}{dcvar_multilevel_fit}(x, ...)

\method{loo}{dcvar_sem_fit}(x, ...)
}
\arguments{
\item{x}{A fitted model object.}

\item{...}{Additional arguments passed to \code{\link[loo:loo]{loo::loo()}}.}
}
\value{
A \code{loo} object from the loo package.
}
\description{
Compute LOO-CV for a fitted model
}
