% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmm_fit-class.R
\name{print.dcVar_hmm_summary}
\alias{print.dcVar_hmm_summary}
\title{Print a dcVar_hmm_summary object}
\usage{
\method{print}{dcVar_hmm_summary}(x, ...)
}
\arguments{
\item{x}{A \code{dcVar_hmm_summary} object as returned by
\code{\link[=summary.dcVar_hmm_fit]{summary.dcVar_hmm_fit()}}.}

\item{...}{Additional arguments (unused).}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print a dcVar_hmm_summary object
}
