% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multilevel_fit-class.R
\name{print.dcVar_multilevel_summary}
\alias{print.dcVar_multilevel_summary}
\title{Print a dcVar_multilevel_summary object}
\usage{
\method{print}{dcVar_multilevel_summary}(x, ...)
}
\arguments{
\item{x}{A \code{dcVar_multilevel_summary} object.}

\item{...}{Additional arguments (unused).}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print a dcVar_multilevel_summary object
}
