% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sem_fit-class.R
\name{print.dcvar_sem_summary}
\alias{print.dcvar_sem_summary}
\title{Print a dcvar_sem_summary object}
\usage{
\method{print}{dcvar_sem_summary}(x, ...)
}
\arguments{
\item{x}{A \code{dcvar_sem_summary} object.}

\item{...}{Additional arguments (unused).}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print a dcvar_sem_summary object
}
