% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcVar_fit-class.R
\name{print.dcVar_summary}
\alias{print.dcVar_summary}
\title{Print a dcVar_summary object}
\usage{
\method{print}{dcVar_summary}(x, ...)
}
\arguments{
\item{x}{A \code{dcVar_summary} object as returned by \code{\link[=summary.dcVar_fit]{summary.dcVar_fit()}}.}

\item{...}{Additional arguments (unused).}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print a dcVar_summary object
}
